backend/requirements.txt
motor
pymongo
fastapi
uvicorn
aiofiles
pydantic
python-dotenv
